requirements.txt
gradio
torch
torchvision
torchaudio
Pillow
imageio
imageio-ffmpeg
numpy
gTTS
opencv-python-headless
